Add output format option capabilities
Adds ability to specify which fields from both q and aux should be output
along with specification of the format per gauge.  Also includes a
refactor of the routines, the addition of a gauge_type, and a few other
changes.  Note that the capability to control these new parameters has
yet to be added and are currently hardcoded.