remove allowflag and many tests from flag2refine2
No longer implicitly define refinement regions when specifying topofiles,
dtopofiles, or qinit files, since this is frequently misleading.
Allow shorter lists to specify these in setrun, or if minlevel,maxlevel
are still specified print a warning that these are no longer used.
Note that logic was repeated in both flag2refine2 and in allowflag to
check these and both sets of tests have been removed.