Fix <= to < in marching_front.py
Most changes are in the docstring, where it incorrectly said that points are
chosen if `Ztopo <= Z2`, for example, when in fact the points are chosen only
if `Ztopo < Z2` when marching.

(Similar updates should be made in clawpack/doc.)

One change in the code: now set `cond0` to initially mark as unchosen those
points for which `Ztopo >= Z2`, rather than `>`, to be consistent with the
the way other points are chosen.

This probably actually have no effect on the points eventually chosen, and
at any rate is a corner case that could only give different behavior
if the topo file being processed has some values exactly equal to `Z2`,
but it is best to make it consistent for clarity.