Fix force_dry checks in filval.f90
Computation of x,y was incorrect; need to account for ghost cells.
Change test from .or. to .and. (time <= time_force_dry).
Checking coarseval(2) == vetac(i,j) wasn't working, maybe need to
reconsider if something like this is needed, which was supposed to be
be applying force_dry_init only to newly created patch cells.
But now it should at least be properly applying only up to time_force_dry.