Add source terms on sphere that have always been missing, turned off for now
For shallow water on the sphere there are additional source terms that
should have been included but have always been missing from GeoClaw.
These are now added to `src/2d/shallow/src2.f90` but are turned off for
now for backward compatibility.

The plan is to turn on the source term in mass by default in v5.10.0,
and in the meantime do more testing and write up a document describing
this change and giving some examples of the impact it has for global scale
modeling. (The source terms in the momentum equations so far seem to make
very little difference and can be dropped for efficiency, but this should
be further investigated.)

To test this currently, change the local variable `sphere_source` and
recompile. The plan for v5.10.0 is to add this variable to
`geoclaw_module.f90` and allow it to be set in `setrun.py`, with default
value 1.