Add support for PyTest
This is an initial go at converting from nose to pytest.  There are a few issues
that were also fixed so that the tests could run to completion that were from
some older files that were not being run.  Note that the regression tests are
not passing barely right now probably due to a slight change to the code in the
intervening time since the tests were working.