New speed_limit parameter to avoid too many dt reductions
A new parameter `rundata.geo_data.speed_limit` can be set in `setrun.py`
and is used in `b4step2.f90` and `getmaxspeed.f90` to scale down the fluid
speed sqrt(u**2 + v**2) if it is larger than this value (preserving the flow
direction).

This helps avoid crashing the code with "too many dt reductions"
in cases when flow off very steep topography with delta B larger
than the fluid depth gives big speeds in the Riemann solution
(shallow water equations aren't valid for flow off a cliff).

If not specified in `setrun.py`, the default value
`speed_limit = 50` m/s (approx 100 knots) is set in `data.py`.
This is much larger than physically reasonable speeds for most problems,
so this should give consistent results with past version of GeoClaw
for runs that don't die along the way.

In practice it is better to use a smaller value for most problems.
Setting `speed_limit = 20` has been found to work well for several
tsunami inundation problems, eliminating the tiny time steps that
are sometimes forced by unphysical huge speeds in one cell, and
allowing jobs to run to completion that otherwise fail in sevaral tests.
This speed is probably larger than fluid speeds expected in most
tsunami, storm surge, or flooding problems.

But more experiments are needed to try to optimize this.  And for some
extreme problems, such as asteroid-impact mega-tsunamis or catastrophic dam
failures, larger values of `speed_limit` may be appropriate.