change dtopotools._load_sift_unit_sources to use pooch to retrieve info_sz.dat
also updated the version in the data/ directory (no change in unit source
data itself) and this will still be used if pooch fails.
`known_hash` can be used in future to see if file has changed online.
This change motivated by the fact that `pip install clawpack` does not
expose the local version in data/ and so using `dtopotools.SiftFault` fails
in Github Actions when building a Juptyer Book there.